Remove 4 unboxed warnings
Use Integer.intValue to explicitly convert to an int

+++ org.eclipse.jgit/src/org/eclipse/jgit/transport/TransportSftp.java
@@ -248,7 +248,8 @@ public class TransportSftp extends SshTransport implements WalkTransport {
 
 				Collections.sort(packs, new Comparator<String>() {
 					public int compare(final String o1, final String o2) {
-						return mtimes.get(o2) - mtimes.get(o1);
+						return mtimes.get(o2).intValue()
+								- mtimes.get(o1).intValue();
 					}
 				});
 			} catch (SftpException je) {
